+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `key.ChartName` for removing workload cluster ID if its present.
+
 ## [6.3.0] - 2022-01-13
 
 ### Added